feat: 完成”人脸库管理“、”学生管理“

+++ src/main/java/com/hzx/conc/entity/Video.java
@@ -23,7 +23,7 @@ public class Video extends BaseEntity {
     //视频名称
     private String vname;
     //上课班级
-    private Integer classId;
+    private Long classId;
     //课程名称
     private String cname;
